# Drop MacroTools dependency