--- a/.github/workflows/CI.yml
+++ b/.github/workflows/CI.yml
@@ -29,7 +29,7 @@ jobs:
           - os: windows-latest
             version: '1'
             arch: x86
-          - os: macos-latest
+          - os: macOS-latest
             version: '1'
             arch: x86
     steps: