--- a/.github/workflows/CI.yml
+++ b/.github/workflows/CI.yml
@@ -29,9 +29,6 @@ jobs:
           - os: windows-latest
             version: '1'
             arch: x86
-          - os: macOS-latest
-            version: '1'
-            arch: x86
     steps:
       - uses: actions/checkout@v2
       - uses: julia-actions/setup-julia@v1