Avoiding the dependency seems like a good idea since it is fairly straightforward to implement here. Also, having this implementation here allows for more optimizations.

Dropping it doesn't seem to affect performance on Julia 1.7.2:

## master

```julia
julia> @time using PrecompileMacro
  0.765824 seconds (611.35 k allocations: 33.541 MiB, 3.44% gc time, 96.21% compilation time)

julia> @time @precompile f() = 3
  0.024938 seconds (33.79 k allocations: 1.939 MiB, 99.43% compilation time)
```

## this PR

```julia
julia> @time using PrecompileMacro
  0.784445 seconds (619.39 k allocations: 33.726 MiB, 96.47% compilation time)

julia> @time @precompile f() = 3
  0.023438 seconds (33.79 k allocations: 1.952 MiB, 99.18% compilation time)
```