@rikhuijzer 2022-02-28
After a few more type annotations and getting rid of a broadcasting, the performance is slightly better:

```julia
julia> @time using PrecompileMacro
  0.739399 seconds (603.59 k allocations: 32.768 MiB, 4.10% gc time, 98.64% compilation time)

julia> @time @precompile f() = 3
  0.023952 seconds (33.79 k allocations: 1.952 MiB, 99.48% compilation time)
```